+++ CHANGELOG.md
@@ -9,6 +9,8 @@ where is hkimura, now?
   willreturn 404. Exception occurs anywhere else?
 
 ## 0.2.5
+### Changed
+* 環境変数 W_I_M で secret の照合をする。
 ### Added
 * simpe authorizing endpoint from where can send current location.
   had better `w.hkim.jp` from smart-phones.